# server/public/storage/.swagger/swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "description": "SwaggerHub API Auto Mocking",
      "url": "https://virtserver.swaggerhub.com/AClain/iTalks/0.4"
    },
    {
      "description": "SwaggerHub API Auto Mocking",
      "url": "http://localhost:8000/api"
    }
  ],
  "info": {
    "description": "iTalks API",
    "version": "0.4",
    "title": "iTalks API",
    "contact": {
      "email": "alexandre.clain11@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Get users",
        "tags": [
          "admins/user"
        ],
        "responses": {
          "201": {
            "description": "Utilisateur enregistré avec succès!"
          }
        }
      },
      "post": {
        "summary": "posts users",
        "tags": [
          "admins/user"
        ],
        "responses": {
          "201": {
            "description": "Le post a été créé avec succès!"
          },
          "500": {
            "description": "Une erreur sest produite, veuillez réessayer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/user/{username}": {
      "post": {
        "summary": "Post user selon username",
        "tags": [
          "admins/user"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "the user username"
          }
        ],
        "responses": {
          "200": {
            "description": "Utilisateur enregistré avec succès!"
          }
        }
      },
      "delete": {
        "summary": "delete user selon username",
        "tags": [
          "admins/user"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "username",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "the user username"
          }
        ],
        "responses": {
          "201": {
            "description": "Utilisateur supprimé avec succès! "
          }
        }
      }
    },
...
    "/image/user/{user_id}/{image_name}": {
      "get": {
        "summary": "Display a User Resource by user_id and image_name",
        "tags": [
          "public"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "User id"
          },
          {
            "in": "path",
            "name": "image_name",
            "required": true,
            "schema": {
              "type": "string",
              "example": "image.jpeg"
            },
            "description": "Resource name"
          }
        ],
        "responses": {
          "201": {
            "description": "Image téléchargé avec succès"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Error returned in case of 4xx HTTP status code",
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "type": "object"
      }
    }
  }
}